requirements.txt
# ── Runtime ────────────────────────────────────────────────────────────────
sqlalchemy>=2.0
pandas>=2.0
numpy>=1.24
faker>=19.0
vaderSentiment>=3.3.2
langchain>=0.2
langchain-core>=0.2
langchain-groq>=0.2
langchain-community>=0.2
streamlit>=1.35
plotly>=5.20
pm4py>=2.7
python-dotenv>=1.0

# ── Development / testing (not installed on Streamlit Cloud) ───────────────
# pip install pytest pytest-cov ruff ipykernel notebook
